transform discordClient to JS Class object
