+++ clients/ai-client.js
@@ -1,9 +1,7 @@
 import OpenAI from 'openai';
 import config from '../config.js';
-import dotenv from 'dotenv';
 
 const openaiKey = config.openaiKey || process.env.OPENAI_API_KEY;
-dotenv.config();
 
 const aiClient = new OpenAI({
   apiKey: openaiKey || process.env.OPENAI_API_KEY,
